pyproject.toml
[tool.poetry]
name = "caopanheiro"
version = "0.1.0"
description = ""
authors = ["Giancarlo Lester <gian.lepear@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
isort = {version = "^5.10.1", allow-prereleases = true}
flake8 = {version = "^4.0.1", allow-prereleases = true}
coverage = "^6.4.1"
parameterized = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | database
  | \.vscode
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 80
skip_gitignore = "true"
skip_glob = ["**/migrations/**", ".venv/**"]
